09 Authorization.

+++ backend/routes/posts.js
@@ -71,7 +71,8 @@ router.post('',
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: url + '/images/' + req.file.filename
+            imagePath: url + '/images/' + req.file.filename,
+            creator: req.userData.userId
         });
         post.save().then(createdPost => {
             res.status(201).json({
@@ -98,13 +99,16 @@ router.put('/:id',
             _id: req.body.id,
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
-        console.log(post);
-        Post.updateOne({_id: req.params.id}, post)
+        Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
             .then(result => {
-                console.log(result);
-                res.status(200).json({message: 'Update successful!'});
+                if (result.nModified > 0) {
+                    res.status(200).json({ message: 'Update successful!' });
+                } else {
+                    res.status(401).json({ message: 'Not authorized!' });
+                }
             });
     }
 );
@@ -112,12 +116,13 @@ router.put('/:id',
 router.delete('/:id',
     checkAuth,
     (req, res, next) => {
-        Post.deleteOne({_id: req.params.id})
+        Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
             .then(result => {
-                console.log(result);
-                res.status(200).json({
-                    message: 'Post deleted!'
-                });
+                if (result.n > 0) {
+                    res.status(200).json({ message: 'Deletion succesful!' });
+                } else {
+                    res.status(401).json({ message: 'Not authorized!' });
+                }
             });
     }
 );
